+++ test/TestAppIsolated/ToolFunctions.cs
@@ -15,6 +15,20 @@ public class TestFunction
         _logger = logger;
     }
 
+    [Function(nameof(HappyFunction))]
+    public string HappyFunction(
+        [McpToolTrigger(nameof(HappyFunction), "Responds to the user with a hello message.")] ToolInvocationContext context,
+        [McpToolProperty(nameof(name), "string", "The name of the person to greet.")] string name,
+        [McpToolProperty(nameof(job), "string", "The job of the person.")] string? job,
+        [McpToolProperty(nameof(age), "number", "The age of the person.")] int age,
+        [McpToolProperty(nameof(isHappy), "boolean", "The happiness of the person.")] bool isHappy
+    )
+    {
+        _logger.LogInformation("C# MCP tool trigger function processed a request: {ToolName}", context.Name);
+        var entityToGreet = name ?? "world";
+        return $"Hello, {entityToGreet}! Job: {job ?? "Unemployed"} | Age: {age} | Happy? {isHappy}.";
+    }
+
     [Function(nameof(GetSnippet))]
     public object GetSnippet(
         [McpToolTrigger(GetSnippetToolName, GetSnippetToolDescription)] ToolInvocationContext context,
@@ -26,13 +40,18 @@ public class TestFunction
     }
 
     [Function(nameof(SaveSnippet))]
-    public void SaveSnippet([McpToolTrigger(SaveSnippetToolName, SaveSnippetToolDescription)] Snippet snippet)
+    public void SaveSnippet(
+        [McpToolTrigger(SaveSnippetToolName, SaveSnippetToolDescription)] Snippet snippet,
+        ToolInvocationContext context)
     {
+        _logger.LogInformation($"Tool name: {context.Name}");
         SnippetsCache.Snippets[snippet.Name] = snippet.Content ?? string.Empty;
     }
 
     [Function(nameof(SearchSnippets))]
-    public object SearchSnippets([McpToolTrigger(SearchSnippetsToolName, SearchSnippetsToolDescription)] SnippetSearchRequest searchRequest)
+    public object SearchSnippets(
+        [McpToolTrigger(SearchSnippetsToolName, SearchSnippetsToolDescription)] SnippetSearchRequest searchRequest,
+        ToolInvocationContext context)
     {
         var comparisonType = searchRequest.CaseSensitive ? StringComparison.Ordinal : StringComparison.OrdinalIgnoreCase;
         return SnippetsCache.Snippets